correction d'erreurs d'incohérences en bdd et appli locale

+++ src/Application/XPControlProduits.java
@@ -10,18 +10,29 @@ public class XPControlProduits {
 	public XPControlProduits(I_Catalogue prods) {
 		produits = prods;
 	}
-	
+
 	public static Boolean XPAjouterProduit(I_Produit p){
-		Boolean out = FrontController.getPDAO().create(p);
-		return out & produits.addProduit(p);
+		if(produits.addProduit(p))
+			return FrontController.getPDAO().create(p);
+		return false;
 	}
 	
 	public static Boolean XPEnleverProduit(String nomProduit){
-		Boolean out = FrontController.getPDAO().delete(nomProduit);
-		return out & produits.removeProduit(nomProduit);
+		if(produits.removeProduit(nomProduit))
+			return FrontController.getPDAO().delete(nomProduit);
+		return false;
 	}
 	
 	public static String[] listeNomsProduits(){
 		return produits.getNomProduits();
+	}
+	
+
+	public static I_Catalogue getProduits() {
+		return produits;
+	}
+
+	public static void setProduits(I_Catalogue produits) {
+		XPControlProduits.produits = produits;
 	}
 }
